rename: 페이지별로 폴더에 분리
