+++ src/js/templates/main/manageItemTemplate.ts
@@ -1,4 +1,4 @@
-import { ItemType } from '../types/types';
+import { ItemType } from '../../types/types';
 
 export const sectionTemplate = {
   inputContainer() {
