+++ src/js/templates/sign/editProfileTemplate.ts
