+++ src/js/templates/sign/signTemplate.ts
