+++ src/js/templates/sign/signUpTemplate.ts
